Ensure that we drop subscriptions exactly once when endpoint starts on a new branch.
It is essential, because itherwise, we may drop
not only inherited, but newly created subscriptions.

We cannot rely only on spec.drop_subscriptions_before_start flag,
because if for some reason compute restarts inside VM,
it will start again with the same spec and flag value.

To handle this, we save the fact of the operation in the database
in the neon_migration_test.drop_subscriptions_done table.
If the table does not exist, we assume that the operation was never performed, so we must do it.
If table exists, we check if the operation was performed on the current timelilne.

Also adjust the test test_subscriber_branching.py to cover this case.